LayoutTests/imported/w3c/web-platform-tests/css/css-inline/baseline-source/baseline-source-first-003-expected.txt




























PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<span data-offset-x="30"></span>
offsetLeft expected 30 but got 10
FAIL .target > * 4 assert_equals:
<div class="inner" style="overflow: hidden; display: inline-block;" data-offset-x="10">
    <div><span></span></div>
    <div><span></span></div>
  </div>
offsetLeft expected 10 but got 20
FAIL .target > * 5 assert_equals:
<span data-offset-x="30"></span>
offsetLeft expected 30 but got 45
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<span data-offset-x="25"></span>
offsetLeft expected 25 but got 10
FAIL .target > * 8 assert_equals:
<div class="inner" style="display: inline-block; columns: 2;" data-offset-x="10">
    <div style="font-size: 20px;"><span></span></div>
    <div style="font-size: 20px;"><span></span></div>
    <div><span></span></div>
    <div><span></span></div>
  </div>
offsetLeft expected 10 but got 53
PASS .target > * 9
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<span data-offset-x="30"></span>
offsetLeft expected 30 but got 45
PASS .target > * 12
FAIL .target > * 13 assert_equals:
<span data-offset-x="30"></span>
offsetLeft expected 30 but got 45
PASS .target > * 14
FAIL .target > * 15 assert_equals:
<span data-offset-x="55"></span>
offsetLeft expected 55 but got 100
PASS .target > * 16
FAIL .target > * 17 assert_equals:
<span data-offset-x="55"></span>
offsetLeft expected 55 but got 57
PASS .target > * 18
FAIL .target > * 19 assert_equals:
<span data-offset-x="70"></span>
offsetLeft expected 70 but got 160
PASS .target > * 20
FAIL .target > * 21 assert_equals:
<span data-offset-x="70"></span>
offsetLeft expected 70 but got 10
FAIL .target > * 22 assert_equals:
<div class="inner" style="display: inline-block;" data-offset-x="10">
    <table>
      <caption></caption>
      <tbody><tr>
        <td style="font-size: 10px;"><span></span><br><span></span></td>
        <td><span></span><br><span></span></td>
      </tr>
      <tr>
        <td><span></span><br><span></span></td>
        <td style="font-size: 10px;"><span></span><br><span></span></td>
      </tr>
      </tbody><caption style="caption-side: bottom;"></caption>
    </table>
  </div>
offsetLeft expected 10 but got 20
